fix(spec-1.7): 配置字段类型 int64→int（libs/config 不支持 int64）+ 加载期守卫测试；dev-verify T19-T21 全过
- AIRUSH_CONSOLE_LLM_DEFAULT_TOKEN_BUDGET 用 int64 声明，libs/config 运行期报"不支持的字段类型"，
  kind 上 migrate Job 与 console 全部起不来——而单测/lint/build 全绿。改 int（Go int 64 位足够）
- cmd/console 加 TestAppConfigAllFieldTypesLoadable：Load[appConfig] 必须成功，把"字段类型不被
  支持"从部署期拉回测试期（已验证：改回 int64 该测试即红）
- 脚本：mac-test-go/mac-vet-integration/mac-lint-go 登记 libs/tenancy libs/llm testkit；
  加 mac-k8s-migrate-debug.sh
- spec-1.7 §11 实施 changelog 6 条（错误码复用、内存基线实测修正、T12 工具调用桥接通过与
  真实供应商探测移交 1.8、testkit 新件、Meter nil 依赖 panic、默认配额消费者）
- CHANGELOG 补 spec-1.7 Added
- dev-verify ALL PASS：T19 无 DB 就绪 / T20 路由+fallback+Responses 桥接+无明文 key / T21 配额面

+++ console/cmd/console/main.go
@@ -43,7 +43,7 @@ type appConfig struct {
 	TSBatchMaxRows int `env:"TS_BATCH_MAX_ROWS" default:"5000"`
 	// spec-1.7：默认租户的月度 LLM token 预算（启动时若无配额行则写入；已有的不覆盖——
 	// 运维经 PUT /api/v1/llm/quota 改过的值优先）。0 = 禁用 LLM。
-	LLMDefaultTokenBudget int64 `env:"LLM_DEFAULT_TOKEN_BUDGET" default:"50000000"`
+	LLMDefaultTokenBudget int `env:"LLM_DEFAULT_TOKEN_BUDGET" default:"50000000"`
 }
 
 // version 由构建期 -ldflags 注入（spec-0.10/0.11 定版链路）。
